requirements.txt
Django~=4.2.14
channels~=4.0.0
asgiref~=3.7.2
djangorestframework~=3.14.0